Add trace classification for Ideation, which we need to understand the differences in performance between QuickCheck and hedgehog.
Just to give an idea, in HH we need 0.22 seconds to generate 2824 traces, and in
QC we need 0.16 seconds to generate 5119 traces (almost twice as many).

Now I have to find out whether CHAIN trace generation is much slower in QC also
due to the size of the traces, blocks, and payload At the moment HH takes 5
seconds to check that only valid traces are generated, whereas QC takes 15. This
might be explained by the fact that QC generates almost twice as many traces. In
CHAIN we use a trace generator inside a trace generator, which might exacerbate
the size difference.

[skip ci]