Fix generation of coverage report
- The coverage reports weren't working because the "doCoverage" haskell.nix
option was not enabled, and the option was referring to the wrong package (was
"decentralized-software-updates" but the package is called
"decentralized-updates").
- Modified the project slightly to more easily toggle generation of coverage
information.
- Commented out the "-Werror" configure flag. This flag was not actually
activated before because it was referring to the wrong package (see above).
After setting it to refer to the correct package, the project no longer builds,
so I've commented it out for the time being. Please review whether or not you
want this flag.